Script V 1.1 - Detect Public IP and generate Private Key for MN

+++ reden-setup.sh
@@ -1,15 +1,12 @@
 #!/bin/bash
 clear
-echo "Make sure you double check before hitting enter! Only one shot at these!"
+if false; then
+echo "REDEN Masternode setup script will detect your Public IP Address\n
+and generate masternode Private Key automatically!\n\n
+Updating system and installing required packages...\n"
+sleep 3
 
-read -t 1 -n 10000 discard 
-read -e -p "Server IP Address : " ip
-read -t 1 -n 10000 discard 
-read -e -p "Masternode Private Key (e.g. 28L11p9KSUQMyw5z6QYay8q68WnNxuH5BbeyAhWutwav1TSNC4S # THE KEY YOU GENERATED EARLIER) : " key
-
-clear
-echo "Updating system and installing required packages..."
-sleep 5
+Port='13058'
 
 # update packages and upgrade Ubuntu
 cd ~
@@ -17,7 +14,7 @@ sudo apt-get -y update
 sudo apt-get -y upgrade
 sudo apt-get dist-upgrade -y
 sudo apt-get -y autoremove
-sudo apt-get -y install wget nano htop git
+sudo apt-get -y install wget nano htop git jq
 sudo apt-get -y install libzmq3-dev
 sudo apt-get -y install libboost-system-dev libboost-filesystem-dev libboost-chrono-dev libboost-program-options-dev libboost-test-dev libboost-thread-dev
 sudo apt-get -y install libevent-dev
@@ -39,28 +36,29 @@ sudo apt-get update -y
 sudo ufw default deny incoming
 sudo ufw default allow outgoing
 sudo ufw allow ssh
-sudo ufw allow 13058/tcp
+sudo ufw allow $Port/tcp
 sudo ufw --force enable
+
 
 #Generating Random Password
 rpcpassword=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
 
 #Find Server Public IP Address
-PublicIP=$(curl ifconfig.co)
-
-Port='13058'
+PublicIP=$(dig +short myip.opendns.com @resolver1.opendns.com)
 
 #Create 2GB swap file
-echo 'Create 2GB disk swap...'
-free -h
-cd /var
-touch swap.img
-chmod 600 swap.img
-dd if=/dev/zero of=/var/swap.img bs=1024k count=2000
-mkswap /var/swap.img
-swapon /var/swap.img
-echo “/var/swap.img none swap sw 0 0” >> /etc/fstab 
-free -h
+if [ ! -f /var/swap.img ]; then
+	echo 'Create 2GB disk swap file...'
+	free -h
+	cd /var
+	touch swap.img
+	chmod 600 swap.img
+	dd if=/dev/zero of=/var/swap.img bs=1024k count=2000
+	mkswap /var/swap.img
+	swapon /var/swap.img
+	echo “/var/swap.img none swap sw 0 0” >> /etc/fstab 
+	free -h
+fi
 
 #Installing Daemon
 cd ~
@@ -70,20 +68,26 @@ cd ~
 #sudo rm reden_ubuntu16_1.0.0_linux.gz
 
 # Copy binaries to /usr/bin
-sudo cp RedenMasternodeSetup/Reden-v1.0-Ubuntu16.04/reden* /usr/bin/
+sudo cp RedenMasternodeSetup/Reden-v1.0-Ubuntu16.04/reden* /usr/bin/ > /dev/null
 
 sudo chmod 755 /usr/bin/reden*
 
 #Stop daemon if it's already running
-reden-cli stop
-sleep 30
+if pgrep -x 'redend' > /dev/null; then
+	reden-cli stop
+	echo 'sleep for 10 seconds...'
+	sleep 10
+fi
 
 #Create reden.conf
-sudo mkdir ~/.redencore
+if [ ! -f ~/.redencore/reden.conf ]; then 
+	sudo mkdir ~/.redencore
+fi
 
+echo 'Creating reden.conf...'
 cat <<EOF > ~/.redencore/reden.conf
 rpcuser=redenrpc
-rpcpassword='$rpcpassword'
+rpcpassword=$rpcpassword
 EOF
 
 sudo chmod 755 -R ~/.redencore/reden.conf
@@ -94,79 +98,151 @@ echo 'sleep for 10 seconds...'
 sleep 10
 
 #Generate masternode private key
+echo 'Generating masternode key...'
 genkey=$(reden-cli masternode genkey)
 reden-cli stop
 
 cat <<EOF > ~/.redencore/reden.conf
 rpcuser=redenrpc
-rpcpassword='$rpcpassword'
+rpcpassword=$rpcpassword
 rpcallowip=127.0.0.1
 listen=1
 server=1
 daemon=1
-logtimestamps=1
 maxconnections=256
-externalip='$PublicIP'
+externalip=$PublicIP
 masternode=1
-masternodeprivkey='$genkey'
+masternodeprivkey=$genkey
 EOF
 
 #Starting daemon second time
 redend
+echo 'sleep for 10 seconds...'
 sleep 10
 
 #Setting auto star for daemon
+echo 'Configuring crontab job...'
+cronjob='@reboot sleep 30 && redenddd'
 crontab -l > tempcron
-echo '@reboot sleep 30 && redend' >> tempcron
-crontab tempcron
+if ! grep -q "$cronjob" tempcron; then
+	echo $cronstr >> tempcron
+	crontab tempcron
+fi
 rm tempcron
 
-cd ~
+#clear
 
-clear
-
-echo "Coin setup complete..."
-echo ""
-echo "Masternode installed with VPS IP Address: $PublicIP"
-echo "Masternode Private Key: $genkey"
-echo "Now, you need to add the following string to the masternode.conf file of your Hot Wallet (the wallet with your Reden funds):"
-echo ""
+echo "========================================================================\n
+Masternode setup complete!\n
+========================================================================\n
+\n
+Masternode was installed with VPS IP Address: $PublicIP\n
+\n
+Masternode Private Key: $genkey\n
+\n
+Now you can add the following string to the masternode.conf file\n
+for your Hot Wallet (the wallet with your Reden collateral funds):\n
+========================================================================\n"
 echo "mn1 $PublicIP:$Port $genkey TxId TxIdx"
-echo ""
-echo "Copy the string above by selecting with mouse followed by Left-Click (Do Not use Ctrl-C) and paste into your masternodes.conf file, then replace mn1 with your desired masternode name TxId with Transaction Id from masternode outputs and TxIdx with Transaction index (0 or 1). Save masternode.conf and restart the wallet."
-echo ""
-sleep 10
-echo "Finally issue a start command for your masternode in the following order:"
-echo "1) Wait for the node wallet on this VPS to sync with other nodes on the network. Eventually the IsSynced status will change to 'true'. It may take from several minutes to hours. Initial status may read '...not in masternode list', which is normal."
-echo "2) Go to your hot wallet and from debug console (Tools->Debug Console) enter:"
-echo ""
-echo "    masternode start-alias <mymnalias>"
-echo ""
-echo "where <mymnalias> is the name of your masternode alias (without brackets) as it was entered in the masternode.conf file."
-echo "once completed please return to this VPS console and wait for the masternode status to change to 'Started'. This will indicate that your masternode is fully functional."
-echo ""
-echo "Your masternode is currently syncing in the background. When you press a key to continue, this message will self-destruct, so please memorize it!"
-echo "The following screen will display current status of this masternode and it's synchronization progress. The data will update in real time every 10 seconds. You can interrupt it at any moment by Ctrl-C."
-echo ""
-echo ""
+echo "========================================================================\n
+\n
+    Use your mouse to copy the whole string above into the clipboard by\n
+    tripple-click + single-click and paste it into your masternodes.conf\n
+    Then replace:\n
+        'mn1' - with your desired masternode name (alias)\n
+        'TxId' - with Transaction Id from masternode outputs\n
+        'TxIdx' - with Transaction Index (0 or 1)\n
+         Remember to save the masternode.conf and restart the wallet!\n
+    \n
+    To introduce your new masternode to the Reden network, you need to
+    issue a masternode start command from your wallet, which proves that\n
+    the collateral for this node is secured.\n
+    \n
+    1) Wait for the node wallet on this VPS to sync with the other nodes\n
+    on the network. Eventually the 'IsSynced' status will change\n
+    to 'true', which will indicate a comlete sync, although it may take\n
+    from several minutes to several hours depending on the network state.\n
+    Your initial Masternode Status may read:\n
+        'Node just started, not yet activated' or\n
+        'Node  is not in masternode list', which is normal and expected.\n
+    \n
+    2) Wait at least until 'IsBlockchainSynced' status becomes 'true'.\n
+    At this point you can go to your wallet and issue a start\n
+    command by either using Debug Console:\n
+        Tools->Debug Console-> enter: masternode start-alias mn1\n
+        where 'mn1' is the name of your masternode (alias)\n
+        as it was entered in the masternode.conf file\n
+        \n
+    or by using wallet GUI:\n
+        Masternodes -> Select masternode -> RightClick -> start alias\n
+    \n
+    Once completed step (2), return to this VPS console and wait for the\n
+    Masternode Status to change to: 'Masternode successfully started'.\n
+    This will indicate that your masternode is fully functional and\n
+    you can celebrate this achievement!\n
+    \n
+    Currently your masternode is syncing with the Reden network...\n
+    Once you press any key to continue, this message will self-destruct!\n
+    Take a moment to re-read it now if anything is not clear...\n
+    \n
+    The following screen will display in real-time the list of\n
+    peer connections, the status of your masternode,\n
+    node synchronization status and additional network and node stats.
+    \n"
+fi
+read -p "*** Press any key to continue ***" -n1 -s
 
-read -p "Press any key to continue... " -n1 -s
-cd ~
+echo "...scroll up to previous screens...\n\n
+Here are some useful commands and tools for troubleshooting:\n
+\n
+========================================================================\n
+To view Redend debug log showing all MN network activity in realtime:\n
+\n
+tail -f ~/.redencore/debug.log \n
+\n
+========================================================================\n
+To view masternode configuration produced by this script in reden.conf:\n
+\n
+cat ~/.redencore/reden.conf
+\n
+NOTE: To edit reden.conf, first stop the redend daemon and start it when finished:
+eden-cli stop\n
+nano ~/.redencore/reden.conf\n
+edend\n
+\n
+========================================================================\n
+To monitor system resource utilization and running processes:\n
+htop\n
+========================================================================\n
+To view in real-time the list of peer connections, the status of your 
+masternode, sync status etc., copy-paste the whole code block below:\n\n
 
-echo ""
-echo "Here are some useful tools and commands for troubleshooting your masternode:"
-echo "(copy/paste without $)"
-echo ""
-echo "Redend debug log showing all MN network activity in real time:"
-echo "$ tail -f ~/.redencore/debug.log"
-echo ""
-echo "To monitor HW and system resource utilization and running processes:"
-echo "$ htop "
-echo ""
-echo "To monitor MN state and its sync status:"
-echo "$ watch -n 10 'reden-cli masternode status && reden-cli mnsync status'"
-echo ""
-echo "If you found this script and MN setup guide helpful, please donate REDEN to: RCdYg5yq3YfymwrZi8EMBSFHxcwR7acniS"
+watch -n 1 \"\
+echo '\nOutbound connections to peer nodes:\n' && \
+reden-cli getpeerinfo | jq -r '.[] | select(.inbound==false) | \
+[.addr, .pingtime, .bytessent, .bytesrecv, .startingheight, \
+.synced_headers, .synced_blocks, .banscore] | @tsv' && \
+echo '\nMasternode Status:' &&  reden-cli masternode status && \
+echo '\nSync Status:' &&  reden-cli mnsync status && \
+echo '\nCurrent Masternode Information:' && reden-cli getinfo && \
+echo '\n\nPress Ctrl-C to Exit...'\"\
+\n\n\
+reden.conf generated by this script located in ~/.redencore directory:\n
+========================================================================"
+cat ~/.redencore/reden.conf
+echo "\
+========================================================================\n
+\n\n
+Thanks for using Reden masternode setup script!
+If you found it helpful, please donate REDEN to:
+RCdYg5yq3YfymwrZi8EMBSFHxcwR7acniS\n
+\n
+Enjoy your Reden Masternode and check back for updates!"
 
-watch -n 10 'reden-cli masternode status && reden-cli mnsync status && reden-cli getinfo'
-
+watch -n 1 "echo 'Outbound connections to peer nodes:\n' && 
+reden-cli getpeerinfo | jq -r '.[] | select(.inbound==false) | \
+[.addr, .pingtime, .bytessent, .bytesrecv, .startingheight, .synced_headers, .synced_blocks, .banscore] | @tsv' && \
+echo '\nMasternode Status:' &&  reden-cli masternode status && \
+echo '\nSync Status:' &&  reden-cli mnsync status && \
+echo '\nCurrent Masternode Information:' && reden-cli getinfo && \
+echo '\n\nPress Ctrl-C to Exit...'"
